+++ Basil/include/Renderer/Framebuffer.h
@@ -14,6 +14,7 @@ namespace Basil
 
 		// Color
 		RGBA8,
+		RED_INTEGER,
 
 		// Depth/stencil
 		DEPTH24STENCIL8,
@@ -59,6 +60,7 @@ namespace Basil
 			virtual void bind() = 0;
 			virtual void unbind() = 0;
 			virtual void resize(uint32_t width, uint32_t height) = 0;
+			virtual int readPixel(uint32_t attachmentIndex, int x, int y) = 0;
 			virtual uint32_t getColorAttachmentRendererID(uint32_t index = 0) const = 0;
 			virtual const FramebufferSpecification& getSpecification() const = 0;
 			static Shared<Framebuffer> create(const FramebufferSpecification& spec);