+++ Basil/src/Platform/OpenGL/OpenGLFramebuffer.cpp
@@ -27,16 +27,16 @@ namespace Basil
 		}
 
 		// Attach a color texture
-		static void attachColorTexture(uint32_t id, int samples, GLenum format, uint32_t width, uint32_t height, int index)
+		static void attachColorTexture(uint32_t id, int samples, GLenum internalFormat, GLenum format, uint32_t width, uint32_t height, int index)
 		{
 			bool multiSampled = samples > 1;
 			if (multiSampled)
 			{
-				glTexImage2DMultisample(GL_TEXTURE_2D_MULTISAMPLE, samples, format, width, height, GL_FALSE);
+				glTexImage2DMultisample(GL_TEXTURE_2D_MULTISAMPLE, samples, internalFormat, width, height, GL_FALSE);
 			}
 			else
 			{
-				glTexImage2D(GL_TEXTURE_2D, 0, format, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
+				glTexImage2D(GL_TEXTURE_2D, 0, internalFormat, width, height, 0, format, GL_UNSIGNED_BYTE, nullptr);
 				glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 				glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 				glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_R, GL_CLAMP_TO_EDGE);
@@ -141,8 +141,10 @@ namespace Basil
 				switch (colorAttachmentSpecifications[i].textureFormat)
 				{
 					case FramebufferTextureFormat::RGBA8:
-						Utils::attachColorTexture(colorAttachments[i], specification.samples, GL_RGBA8, specification.width, specification.height, i);
+						Utils::attachColorTexture(colorAttachments[i], specification.samples, GL_RGBA8, GL_RGBA, specification.width, specification.height, i);
 						break;
+					case FramebufferTextureFormat::RED_INTEGER:
+						Utils::attachColorTexture(colorAttachments[i], specification.samples, GL_R32I, GL_RED_INTEGER, specification.width, specification.height, i);
 				}
 			}
 		}
@@ -200,6 +202,17 @@ namespace Basil
 		specification.width = width;
 		specification.height = height;
 		invalidate();
+	}
+
+	// Read pixel entity value
+	int OpenGLFramebuffer::readPixel(uint32_t attachmentIndex, int x, int y)
+	{
+		ASSERT(attachmentIndex < colorAttachments.size(), "");
+
+		glReadBuffer(GL_COLOR_ATTACHMENT0 + attachmentIndex);
+		int pixelData;
+		glReadPixels(x, y, 1, 1, GL_RED_INTEGER, GL_INT, &pixelData);
+		return pixelData;
 	}
 
 	// Return the color attachment renderer ID