+++ Pesto/include/EditorLayer.h
@@ -35,6 +35,8 @@ namespace Basil
 			Shared<Scene> activeScene;
 
 			glm::vec2 viewportSize = { 0.0f, 0.0f };
+			glm::vec2 viewportBounds[2];
+
 			glm::vec4 squareColor = { 0.2f, 0.3f, 0.8f, 1.0f };
 			
 			EditorCamera editorCamera;